engine: extract esm/layers/table.js
